SONAR-6831 CeWorkerRunnable is now an interface and impl class in pico
